feat(1560): cross-adapter venue-weekend campout heuristic
PR H closes audit bucket A.2 (#1718) — events with informal
`Friday:/Saturday:/Sunday:` section labels in description that
weren't being recognized as multi-day spans.

The heuristic was already correct in Hash Rego's parser (PR #1637)
but lived in the adapter, so non-Hash-Rego events (Hollyweird HapPy
Hour Brewfish/Tin Fish, RIH3 40th Analversary, KWH3 Anniversary 1993)
never had it run. Lift it into the merge pipeline so it generalizes:

1. Extract `detectVenueWeekendEndDate` + supporting weekday allow-list /
   trigger regex into `src/pipeline/venue-weekend.ts` (104 LOC).
2. Re-export from `src/adapters/hashrego/parser.ts` for backward
   compatibility — the existing parser test suite + the adapter's
   per-day kennel-attribution branch still call it the same way.
3. Wire it into `sanitizeRawFields` in `src/pipeline/merge.ts` so it
   runs on every RawEventData. Gates:
     - `!event.endDate` — don't override an explicit source range
     - `!event.seriesId && !event.seriesParent` — series logic owns
       endDate for parent/child rows separately
     - `event.description && event.date` — need both to anchor on
   Hash Rego is a no-op: the parser already populated endDate before
   the RawEventData reached merge.

Live-verified against the Madison Token Run Campout page — parser
still emits endDate=2026-05-31 through the re-export. 6 new unit
tests in `src/pipeline/venue-weekend.test.ts` cover the audit anchor
cases + the MAX_FORWARD_OFFSET cap.

Stale Hash Rego canonical rows (Madison, Oregon HHH) will populate
endDate on their next scrape via the existing parser path. The 4
non-Hash-Rego events surface endDate on next scrape via the new
pipeline-level call.

Refs #1560, #1718.
